l02_2_tool_calling_langchain/requirements.txt
langchain
langchain-google-vertexai
python-whois
